Create N-Queens.cpp
o, basically we have been given n queens and we need to place those queens in the board such that no two queens attack each other right?

Cool, so we know that the number of ways to place N queens on a NxN board can get very large since we have N^2 choices at first, then N^2 -1, N^2 -2 and so on... which eventually leads to an overall time complexity of O(N^2N), which is extremely high and we can't afford that.

But, simultaneously we also know that we don't actually need to traverse all the O(N^2) options each time.
So, we can place one queen in each row and then move to the next row.

So for each row, we will place one queen (if possible within the contraint) and then move on to the next row. This will be repeated till all N queens have been placed. We will use the isSafe method to ensure that the queen is safe to place at (i, j). Once, all N queens have been placed, we have got our first solution.

Finally, to get all the other possible solutions, we will keep on removing the previously placed queen and try if its possible to place it at some other location, i.e, where we need to backtrack. This will give us all the possible combinations to place N Queens on the board as per the given constraints.